SAM-14, Вынес расширение UISlider в отдельный файл
