+++ PhotoTune/Modules/EditingScreen/Views/Tools/Tune/ToolSlider.swift
@@ -81,23 +81,3 @@ final class ToolSlider: UISlider
 		fatalError("init(coder:) has not been implemented")
 	}
 }
-
-extension UISlider
-{
-	var trackBounds: CGRect {
-		return trackRect(forBounds: bounds)
-	}
-
-	var trackFrame: CGRect {
-		guard let superView = superview else { return CGRect.zero }
-		return self.convert(trackBounds, to: superView)
-	}
-
-	var thumbBounds: CGRect {
-		return thumbRect(forBounds: frame, trackRect: trackBounds, value: value)
-	}
-
-	var thumbFrame: CGRect {
-		return thumbRect(forBounds: bounds, trackRect: trackFrame, value: value)
-	}
-}
